[SPARK-36821][SQL] Make class ColumnarBatch extendable - addendum
### What changes were proposed in this pull request?
A follow up of https://github.com/apache/spark/pull/34054. Three things changed:
1. Add a test for extendable class `ColumnarBatch`
2. Make `ColumnarBatchRow` public.
3. Change private fields to protected fields.

### Why are the changes needed?
A follow up of https://github.com/apache/spark/pull/34054. Class ColumnarBatch need to be extendable to support better vectorized reading in multiple data sources. For example, Iceberg needs to filter out deleted rows in a batch before Spark consumes it, to support row-level delete( apache/iceberg#3141) in vectorized read.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
A new test is added

Closes #34087 from flyrain/SPARK-36821.

Authored-by: Yufei Gu <yufei_gu@apple.com>
